@@ -1,14 +1,26 @@
-using Microsoft.AspNetCore;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 
-namespace DNVGL.SolutionPackage.Demo
+namespace DNVGL.OAuth.Demo
 {
 	public class Program
 	{
 		public static void Main(string[] args)
 		{
-			WebHost.CreateDefaultBuilder(args).UseStartup<Startup>().Build().Run();
+			CreateHostBuilder(args).Build().Run();
 		}
+
+		public static IHostBuilder CreateHostBuilder(string[] args) =>
+			Host.CreateDefaultBuilder(args)
+				.ConfigureWebHostDefaults(webBuilder =>
+				{
+					webBuilder.UseStartup<Startup>();
+				});
 	}
 }
